% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{get_days_in_year}
\alias{get_days_in_year}
\title{Return the number of days for the given year.}
\usage{
get_days_in_year(given_year)
}
\arguments{
\item{given_year}{year for which to return data.}
}
\value{
number of days in that year.
}
\description{
Return the number of days for the given year.
}
\examples{
# Example
get_days_in_year(2024)
}
