% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medacoPlot-package.R
\docType{package}
\name{medacoPlot-package}
\alias{medacoPlot}
\alias{medacoPlot-package}
\title{medacoPlot: Renders plots of medaco csv data}
\description{
Plot power input and output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://joergsesterhenn.github.io/medaco-plot/}
}

}
\author{
\strong{Maintainer}: Joerg Sesterhenn \email{joerg.sesterhenn@gmail.com}

}
\keyword{internal}
