% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Generic Plot Function}
\usage{
plot(plot_type, data)
}
\arguments{
\item{plot_type}{Character, the plot type from the dropdown menu.}

\item{data}{Data frame containing data for plotting.}
}
\value{
A ggplot object created by the appropriate plotting function.
}
\description{
Selects a plot function based on the specified plot type and dataset.
}
\examples{
plot("by month", data)
}
