% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_aggregated_by_hour}
\alias{plot_aggregated_by_hour}
\title{Plot Aggregated Data by Hour}
\usage{
plot_aggregated_by_hour(df)
}
\arguments{
\item{df}{Data frame containing hourly data in long format.}
}
\value{
A ggplot object showing hourly aggregated values.
}
\description{
Generates a bar plot showing hourly input and output sums.
}
\examples{
plot_aggregated_by_hour(df)
}
