% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_aggregated_by_month}
\alias{plot_aggregated_by_month}
\title{Plot Aggregated Data by Month}
\usage{
plot_aggregated_by_month(df)
}
\arguments{
\item{df}{Data frame containing monthly data in long format.}
}
\value{
A ggplot object showing monthly aggregated values.
}
\description{
Generates a bar plot showing monthly input and output sums.
}
\examples{
plot_aggregated_by_month(df)
}
