% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_by_hour_and_month}
\alias{plot_by_hour_and_month}
\title{Plot Hourly Data by Month}
\usage{
plot_by_hour_and_month(df)
}
\arguments{
\item{df}{Data frame containing hourly data by month in long format.}
}
\value{
A ggplot object showing hourly values by month in facets.
}
\description{
Generates a bar plot showing hourly input and output sums by month.
}
\examples{
plot_by_hour_and_month(df)
}
