% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Generate a Heatmap of Hourly Data by Month}
\usage{
plot_heatmap(df)
}
\arguments{
\item{df}{Data frame containing hourly data by month in long format.}
}
\value{
A ggplot object with a heatmap representing input/output.
}
\description{
Creates a heatmap to show hourly input and output data across months.
}
\examples{
plot_heatmap(df)
}
