% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_line_chart}
\alias{plot_line_chart}
\title{Generate a Line Chart of Hourly Data by Month}
\usage{
plot_line_chart(df)
}
\arguments{
\item{df}{Data frame containing hourly data by month in long format.}
}
\value{
A ggplot object with a line chart.
}
\description{
Creates a line chart to show input/output data by hour for each month.
}
\examples{
plot_line_chart(df)
}
