% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{plot_map}
\alias{plot_map}
\title{Map of Dropdown Items to Plot Functions}
\format{
A data frame with dropdown items as row names and plot function
names as column values.
}
\usage{
plot_map
}
\description{
A data frame mapping dropdown box items to corresponding plot functions.
}
\keyword{datasets}
