% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ridgeline}
\alias{plot_ridgeline}
\title{Generate a Ridgeline Plot of Hourly Data by Month}
\usage{
plot_ridgeline(df)
}
\arguments{
\item{df}{Data frame containing hourly data by month in long format.}
}
\value{
A ggplot object with a ridgeline plot.
}
\description{
Creates a ridgeline plot to show distribution of input/output data by month.
}
\examples{
plot_ridgeline(df)
}
