% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stacked_area}
\alias{plot_stacked_area}
\title{Generate a Stacked Area Chart of Hourly Data by Month}
\usage{
plot_stacked_area(df)
}
\arguments{
\item{df}{Data frame containing hourly data by month in long format.}
}
\value{
A ggplot object with a stacked area chart.
}
\description{
Creates a stacked area chart to visualize input/output data by month.
}
\examples{
plot_stacked_area(df)
}
