% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_power_data}
\alias{read_power_data}
\title{Read Power Data}
\usage{
read_power_data(data_path)
}
\arguments{
\item{data_path}{Character string specifying the directory containing
\code{.CSV} files to read.}
}
\value{
A data frame containing timestamp, input, and output columns
from the files.
}
\description{
Reads power data from \code{.CSV} files in the specified directory,
selecting specific columns and using a custom locale for decimal
and grouping marks.
}
\examples{
read_power_data("data/")
}
